Convert `VerificationError` into `VerificationResultMetadata`
The `VerificationError` has been changed to
`VerificationResultMetadata`. This change was made to better capture the
intent that it is intended to store the values used during verification.
While the `VerificationResult` is meant to communicate success or
failure. This allows the metadata to be used for displaying what
happened in the success branches of the verification process.