Break `VerificationResultMetadat` into unique types
Previously `VerificationResultMetadat` was an enumeration of all the
potential failures. This resulted in needing to have match statements
for each potential variant and making it more work to add new variants.
Now each type that implements a `VerificationMessage` is unique and the
`Verifier::verify()` is generic over these types.

`ResultMessage` has been renamed to `VerificationMessage`.
The use of `Metadata` in types that implement `VerificationMessage` has
been changed to `Values`. i.e. `ReportDataVerificationMetadata` ->
`ReportDataVerificationValues`