Add logic to decode and verify certificate (#49)
* Add logic to decode and verify certificate

A DER encoded certificate can now be decoded and have it's signature
verified

* Add alloc feature

Previously the x509 certificate logic was always built as part of the
`mc-attestation-verifier` crate. Now the x509 certificate logic is
guarded by the `alloc` feature so that the crate can still be built in
no alloc environments with limited capability.

* Move test PEM contents to separate files

Previously the PEM strings used in testing were embedded in the test
modules. Now the pem contents are in individual files which are accessed
via `include_str!`.