Add verification of certificate time (#51)
* Add verification of certificate time

Previously the certificate time was not checked to be within range of
the `Validity`, https://www.rfc-editor.org/rfc/rfc5280#section-4.1.2.5.
Now the certificate time is verified to be within the `Validity`

* Add `CertificateNotYetValid` error

Previously if a certificate was not yet valid it would return a
`CertificateExpired` error. Now a dedicated `CertificateNotYetValid`
error is returned.