Fix logic to update README of anchor crate.
When using `cargo release` one crate in a repository has settings on
how to update the CHANGELOG. This is the anchor crate. Since this crate
has the Cargo.toml entry `[package.metadata.release]` it will not look
at the entry in the root `Cargo.toml`. The common settings of updating
the README must be repeated in the anchor crates
`[package.metadata.release]`.