Add `Serialize` and `Deserialize` to `Error`
Any error that is meant out be propagated out of an enclave needs to
be able to be serialized and deserialized from the trusted side into the
untrusted side.

`Clone` was also added as the changes needed for `Serialize` and
`Deserialize` made it possible for the error to be cloneable.